feat: implement domain-driven design principles across authentication and blog modules; refactor command handlers and use cases to utilize Result type for consistent error handling and improve overall code clarity

+++ src/modules/blog/application/use-cases/delete-blog.use-case.ts
@@ -1,14 +1,28 @@
 import { Injectable } from '@nestjs/common';
 
+import { DomainExceptionCode } from '@/core/exceptions/domain-exception-code.enum';
+import { DomainException } from '@/core/exceptions/domain.exception';
+import { Result } from '@/core/result/result.factory';
+import { Result as ResultType } from '@/core/result/result.types';
 import { IUseCase } from '@/core/types/use-case';
 
 import { BlogRepository } from '../../infrastructure/blog.repository';
 
 @Injectable()
-export class DeleteBlogUseCase implements IUseCase<string, boolean> {
+export class DeleteBlogUseCase implements IUseCase<string, ResultType<null>> {
   constructor(private readonly blogRepository: BlogRepository) {}
 
-  async execute(id: string): Promise<boolean> {
-    return this.blogRepository.deleteBlogById(id);
+  async execute(id: string): Promise<ResultType<null>> {
+    const blog = await this.blogRepository.findById(id);
+    if (!blog) {
+      return Result.fail(DomainExceptionCode.NotFound);
+    }
+
+    const deleteResult = await this.blogRepository.deleteBlogById(id);
+    if (!deleteResult) {
+      throw new DomainException(DomainExceptionCode.InternalServerError);
+    }
+
+    return Result.ok(null);
   }
 }
